ReFact: AddReservation,AuthToken,RefreshToken
Changing the refresh token and delete token function. When you first log in and log out, you are directed to the staff login page. The token that is deleted is only the staff token, not the owner token, so there will be no case of not being able to log in because the token is not there.